Move character results to bottom
The original location above stages looks good, but if you play as or against many characters it splits the overall results very far from the rest of the concise results. It would also make it weird to compare new results to screenshots from old versions.

There are better solutions but I think this is good for now.